Add files via upload
A wrapper for the tool minicom. This tool writes the capture files to a specific directory. Say you need to console to a machine that is of a certain time and you wish to keep that separated from the rest, you could use this wrapper. 

$ console.py -p AMD umpalumpa

This would create a directory with the name 'AMD' and that directory would contain a file with the name umpalumpa_2022-03-28_16:17:13.capture.